# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/api/documents/search": {
      "get": {
        "tags": [
          "documents"
        ],
        "summary": "A002:Search Document",
        "description": "指定したクエリをもとにドキュメントを検索します。\n\nArgs:\n    user(User): 認証済みユーザー情報\n    service(DocumentService): ドキュメントサービス\n    q(str): 検索クエリ\n    hits_per_page(int): 1ページに含めるデータの最大数(デフォルト:7)\n    page(int): データを取得したいページ番号(デフォルト:1)\n    uploader(list[Literal['user', 'admin']]): ドキュメントのアップロード者('user'または'admin')\n    genre(str | None): AUTOSARドキュメントのワーキンググループ\n    release(str | None): AUTOSARドキュメントのリリースバージョン\n\nReturns:\n    SearchDocumentsResponse: 検索結果をマッピングしたレスポンス。\n\nRaises:\n    HTTPException(400): クエリパラメータ関係のエラーが発生した場合。\n    HTTPException(500): サーバーエラーが発生した場合。",
        "operationId": "A002",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "検索クエリ",
              "examples": [
                "CAN driver"
              ],
              "title": "Q"
            },
            "description": "検索クエリ"
          },
          {
            "name": "hits_per_page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "1ページに含める検索結果の数",
              "examples": [
                7
              ],
              "default": 7,
              "title": "Hits Per Page"
            },
            "description": "1ページに含める検索結果の数"
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "description": "取得したいページ番号",
              "examples": [
                1
              ],
              "default": 1,
              "title": "Page"
            },
            "description": "取得したいページ番号"
          },
          {
            "name": "uploader[]",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "enum": [
                      "user",
                      "admin"
                    ],
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "ドキュメントのアップロード者(['user']または['admin']または['user', 'admin']のリスト)",
              "examples": [
                [
                  "admin"
                ]
              ],
              "default": [
                "user",
                "admin"
              ],
              "title": "Uploader[]"
            },
            "description": "ドキュメントのアップロード者(['user']または['admin']または['user', 'admin']のリスト)"
          },
          {
            "name": "genre",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "AUTOSARドキュメントのワーキンググループ",
              "examples": [
                "SWS,EXP"
              ],
              "title": "Genre"
            },
            "description": "AUTOSARドキュメントのワーキンググループ"
          },
          {
            "name": "release",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "AUTOSARドキュメントのリリースバージョン",
              "examples": [
                "R4-2-2"
              ],
              "title": "Release"
            },
            "description": "AUTOSARドキュメントのリリースバージョン"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchDocumentsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "missing_token": {
                    "summary": "Missing token",
                    "value": {
                      "detail": "Authorization header missing"
                    }
                  },
                  "invalid_or_expired_token": {
                    "summary": "Invalid or expired token",
                    "value": {
                      "detail": "Invalid or expired token"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "required": [
                    "detail"
                  ]
                },
                "example": {
                  "detail": "Internal Server Error"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "invalid_query": {
                    "value": {
                      "detail": "Invalid query parameter value"
                    }
                  }
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/upload": {
      "post": {
        "tags": [
          "documents"
        ],
        "summary": "A016:Upload Document",
        "description": "ファイルアップロードエンドポイント\n\nArgs:\n    file (UploadFile): アップロードされたファイル\n    user (User): 現在のユーザー\n    service (DocumentService): ドキュメントサービス\n    s3_client (S3Client): S3クライアント\n\nReturns:\n    DocumentBase: アップロードされたドキュメント情報\n\nRaises:\n    HTTPException: 処理失敗時のエラー",
        "operationId": "A016",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_A016"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentBase"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "missing_token": {
                    "summary": "Missing token",
                    "value": {
                      "detail": "Authorization header missing"
                    }
                  },
                  "invalid_or_expired_token": {
                    "summary": "Invalid or expired token",
                    "value": {
                      "detail": "Invalid or expired token"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ]
                },
                "example": {
                  "detail": "Internal Server Error"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "type": "object",
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "invalid_file": {
                    "value": {
                      "detail": "Invalid file request"
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/documents/{document_id}": {
      "delete": {
        "tags": [
          "documents"
        ],
        "summary": "A017:Delete Document",
        "description": "指定されたドキュメントを削除する\n\nMeilisearchのドキュメント削除に失敗した場合は検索APIやエージェントツールで使用\nされてしまうため以下の順番で削除処理を行い、途中で失敗した場合は後続処理へ進まない\n    1. Meilisearch\n    2. DB\n    3. S3\n\nArgs:\n    document_id: 削除対象のドキュメントのID\n    user: リクエストしたユーザー情報\n    service: ドキュメントサービス\n    s3_client: AWS S3のクライアント\n\nReturns:\n    Response: ステータスコード204で削除が成功したことを示す\n\nRaises:\n    HTTPException:\n        - 403: ユーザーに削除する権限がないとき\n        - 404: 指定されたドキュメントが存在しないとき",
        "operationId": "A017",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "ドキュメントID",
              "examples": [
                "260ea4dd-36a7-401f-b2a1-d6a28d97f140"
              ],
              "title": "Document Id"
            },
            "description": "ドキュメントID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "missing_token": {
                    "summary": "Missing token",
                    "value": {
                      "detail": "Authorization header missing"
                    }
                  },
                  "invalid_or_expired_token": {
                    "summary": "Invalid or expired token",
                    "value": {
                      "detail": "Invalid or expired token"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "required": [
                    "detail"
                  ]
                },
                "example": {
                  "detail": "Internal Server Error"
                }
              }
            }
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "not_authorized": {
                    "value": {
                      "detail": "Not authorized to delete this document"
                    }
                  }
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "detail": {
                      "type": "string",
                      "description": "Error detail message"
                    }
                  },
                  "required": [
                    "detail"
                  ]
                },
                "examples": {
                  "not_found": {
                    "value": {
                      "detail": "Document not found"
                    }
                  }
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_A016": {
        "properties": {
          "file": {
            "type": "string",
            "format": "binary",
            "title": "File",
            "description": "ユーザーがアップロードしたファイル"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_A016"
      },
...
      "DocumentBase": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "ドキュメントID",
            "examples": [
              "260ea4dd-36a7-401f-b2a1-d6a28d97f140"
            ]
          },
          "title": {
            "type": "string",
            "title": "Title",
            "description": "ドキュメントのタイトル",
            "examples": [
              "Specification of CAN Driver"
            ]
          },
          "path": {
            "type": "string",
            "title": "Path",
            "description": "ドキュメントのパス",
            "examples": [
              "/app/public/documents/R22-11/AUTOSAR_SWS_CANDriver.pdf"
            ]
          },
          "subject": {
            "type": "string",
            "enum": [
              "AUTOSAR",
              "others"
            ],
            "title": "Subject",
            "description": "ドキュメントの種類",
            "examples": [
              "AUTOSAR"
            ]
          },
          "genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre",
            "description": "AUTOSARドキュメントのワーキンググループ",
            "examples": [
              "SWS"
            ]
          },
          "release": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release",
            "description": "AUTOSARドキュメントのリリースバージョン",
            "examples": [
              "R22-11"
            ]
          },
          "file_type": {
            "type": "string",
            "title": "File Type",
            "description": "ドキュメントのMIMEタイプ",
            "examples": [
              "application/pdf"
            ]
          },
          "summary": {
            "type": "string",
            "title": "Summary",
            "description": "ドキュメントの要約",
            "default": "",
            "examples": [
              "This document describes the CAN driver specification..."
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "path",
          "subject",
          "file_type"
        ],
        "title": "DocumentBase",
        "description": "ドキュメントの基本情報"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SearchDocumentsResponse": {
        "properties": {
          "total_pages": {
            "type": "integer",
            "title": "Total Pages",
            "description": "検索結果のページ総数",
            "examples": [
              3
            ]
          },
          "documents": {
            "items": {
              "$ref": "#/components/schemas/DocumentBase"
            },
            "type": "array",
            "title": "Documents",
            "description": "検索結果のリスト",
            "examples": [
              [
                {
                  "file_type": "application/pdf",
                  "genre": "SWS",
                  "id": "260ea4dd-36a7-401f-b2a1-d6a28d97f140",
                  "path": "/app/public/documents/R22-11/AUTOSAR_SWS_CANDriver.pdf",
                  "release": "R22-11",
                  "subject": "AUTOSAR",
                  "title": "Specification of CAN Driver"
                }
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "total_pages",
          "documents"
        ],
        "title": "SearchDocumentsResponse",
        "description": "ドキュメント検索レスポンス"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}